requirements.txt
Flask
Flask-WTF
Flask-Uploads
python-dotenv
openai
geopy